+++ NEWS.md
@@ -1,4 +1,5 @@
 # cancensus - 0.5.3
+- Added a check and context menu to install `sf` package when user requests spatial data but does not have the required package installed as opposed to erroring out. 
 - fixes a bug in the local data recall check
 
 # cancensus - 0.5.2